Move main window views beside window
