+++ SentryReplay/AboutView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="SentryReplay.Controls.AboutHelpView"
+﻿<UserControl x:Class="SentryReplay.AboutView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
