+++ SentryReplay/AppResources.xaml
